+++ persistent-mysql/test/main.hs
@@ -50,6 +50,7 @@ import qualified TransactionLevelTest
 import qualified UniqueTest
 import qualified UpsertTest
 import qualified CustomConstraintTest
+import qualified LongIdentifierTest
 
 type Tuple a b = (a, b)
 
@@ -125,6 +126,7 @@ main = do
       , CustomPrimaryKeyReferenceTest.migration
       , MigrationColumnLengthTest.migration
       , TransactionLevelTest.migration
+      -- , LongIdentifierTest.migration
       ]
     PersistentTest.cleanDB
 
@@ -188,6 +190,9 @@ main = do
 
     MigrationIdempotencyTest.specsWith db
     CustomConstraintTest.specs db
+    -- TODO: implement automatic truncation for too long foreign keys, so we can run this test.
+    xdescribe "The migration for this test currently fails because of MySQL's 64 character limit for identifiers. See https://github.com/yesodweb/persistent/issues/1000 for details" $
+        LongIdentifierTest.specsWith db
 
 roundFn :: RealFrac a => a -> Integer
 roundFn = round