+++ persistent-postgresql/Database/Persist/Postgresql.hs
@@ -566,15 +566,13 @@ migrate' allDefs getter entity = fmap (fmap $ map showAlterDb) $ do
         (errs, _) -> return $ Left errs
   where
     name = entityDB entity
-    (newcols', udefs, fdefs) = mkColumns allDefs entity
+    (newcols', udefs, fdefs) = postgresMkColumns allDefs entity
     migrationText exists old''
         | not exists =
             createText newcols fdefs udspair
         | otherwise =
             let (acs, ats) =
-                    getAlters allDefs entity (newcols, udspair)
-                    -- $ excludeForeignKeys
-                    $ old'
+                    getAlters allDefs entity (newcols, udspair) old'
                 acs' = map (AlterColumn name) acs
                 ats' = map (AlterTable name) ats
             in
@@ -583,13 +581,6 @@ migrate' allDefs getter entity = fmap (fmap $ map showAlterDb) $ do
          old' = partitionEithers old''
          newcols = filter (not . safeToRemove entity . cName) newcols'
          udspair = map udToPair udefs
-         excludeForeignKeys (xs,ys) = (map excludeForeignKey xs,ys)
-         excludeForeignKey c = case cReference c of
-           Just (_,fk) ->
-             case find (\f -> fk == foreignConstraintNameDBName f) fdefs of
-               Just _ -> c { cReference = Nothing }
-               Nothing -> c
-           Nothing -> c
             -- Check for table existence if there are no columns, workaround
             -- for https://github.com/yesodweb/persistent/issues/152
 
@@ -1223,6 +1214,38 @@ instance PersistConfig PostgresConf where
         addPass     = maybeAddParam "password" "PGPASS"
         addDatabase = maybeAddParam "dbname"   "PGDATABASE"
 
+refName :: DBName -> DBName -> DBName
+refName (DBName table) (DBName column) =
+    let overhead = T.length $ T.concat ["_", "_fkey"]
+        (fromTable, fromColumn) = shortenNames overhead (T.length table, T.length column)
+    in DBName $ T.concat [T.take fromTable table, "_", T.take fromColumn column, "_fkey"]
+
+    where
+
+      -- Postgres automatically truncates too long foreign keys to a combination of
+      -- truncatedTableName + "_" + truncatedColumnName + "_fkey"
+      -- This works fine for normal use cases, but it creates an issue for Persistent
+      -- Because after running the migrations, Persistent sees the truncated foreign key constraint
+      -- doesn't have the expected name, and suggests that you migrate again
+      -- To workaround this, we copy the Postgres truncation approach before sending foreign key constraints to it.
+      --
+      -- I believe this will also be an issue for extremely long table names,
+      -- but it's just much more likely to exist with foreign key constraints because they're usually tablename * 2 in length
+
+      -- Approximation of the algorithm Postgres uses to truncate identifiers
+      -- See makeObjectName https://github.com/postgres/postgres/blob/5406513e997f5ee9de79d4076ae91c04af0c52f6/src/backend/commands/indexcmds.c#L2074-L2080
+      shortenNames :: Int -> (Int, Int) -> (Int, Int)
+      shortenNames overhead (x, y)
+           | x + y + overhead <= maximumIdentifierLength = (x, y)
+           | x > y = shortenNames overhead (x - 1, y)
+           | otherwise = shortenNames overhead (x, y - 1)
+
+-- | Postgres' default maximum identifier length in bytes
+-- (You can re-compile Postgres with a new limit, but I'm assuming that virtually noone does this).
+-- See https://www.postgresql.org/docs/11/sql-syntax-lexical.html#SQL-SYNTAX-IDENTIFIERS
+maximumIdentifierLength :: Int
+maximumIdentifierLength = 63
+
 udToPair :: UniqueDef -> (DBName, [DBName])
 udToPair ud = (uniqueDBName ud, map snd $ uniqueFields ud)
 
@@ -1245,7 +1268,7 @@ mockMigrate allDefs _ entity = fmap (fmap $ map showAlterDb) $ do
                  in  acs' ++ ats'
        where
          old' = partitionEithers old''
-         (newcols', udefs, fdefs) = mkColumns allDefs entity
+         (newcols', udefs, fdefs) = postgresMkColumns allDefs entity
          newcols = filter (not . safeToRemove entity . cName) newcols'
          udspair = map udToPair udefs
             -- Check for table existence if there are no columns, workaround
@@ -1345,3 +1368,7 @@ migrateEnableExtension extName = WriterT $ WriterT $ do
   if res == [Single 0]
     then return (((), []) , [(False, "CREATe EXTENSION \"" <> extName <> "\"")])
     else return (((), []), [])
+
+postgresMkColumns :: [EntityDef] -> EntityDef -> ([Column], [UniqueDef], [ForeignDef])
+postgresMkColumns allDefs t = mkColumns allDefs t (emptyBackendSpecificOverrides {backendSpecificForeignKeyName = Just refName})
+