+++ persistent-postgresql/test/main.hs
@@ -52,6 +52,7 @@ import qualified TreeTest
 import qualified UniqueTest
 import qualified UpsertTest
 import qualified CustomConstraintTest
+import qualified LongIdentifierTest
 
 type Tuple = (,)
 
@@ -121,6 +122,7 @@ main = do
       , CustomPrimaryKeyReferenceTest.migration
       , MigrationColumnLengthTest.migration
       , TransactionLevelTest.migration
+      , LongIdentifierTest.migration
       , ForeignKey.compositeMigrate
       ]
     PersistentTest.cleanDB
@@ -184,5 +186,6 @@ main = do
     TransactionLevelTest.specsWith db
     JSONTest.specs
     CustomConstraintTest.specs db
+    LongIdentifierTest.specsWith db
     -- FIXME: not used, probably should?
     -- ArrayAggTest.specs db