+++ persistent-sqlite/test/main.hs
@@ -41,6 +41,7 @@ import qualified SumTypeTest
 import qualified TransactionLevelTest
 import qualified UniqueTest
 import qualified UpsertTest
+import qualified LongIdentifierTest
 
 import Control.Exception (handle, IOException)
 import Control.Monad.Catch (catch)
@@ -153,6 +154,7 @@ main = do
       , CustomPrimaryKeyReferenceTest.migration
       , MigrationColumnLengthTest.migration
       , TransactionLevelTest.migration
+      , LongIdentifierTest.migration
       ]
     PersistentTest.cleanDB
 
@@ -213,6 +215,7 @@ main = do
     ForeignKey.specsWith db
     TransactionLevelTest.specsWith db
     MigrationTest.specsWith db
+    LongIdentifierTest.specsWith db
 
     it "issue #328" $ asIO $ runSqliteInfo (mkSqliteConnectionInfo ":memory:") $ do
         runMigrationSilent migrateAll